added domain package and moved Offer.java and OfferCriteria.java
